refactor(app): extract media support
